requirements-test.txt
# Testing dependencies for ACWD Home Assistant Integration
pytest-homeassistant-custom-component>=0.13.310
pytest-mock>=3.10.0
freezegun>=1.2.0
